hard to use templated solvers for specific cases
