+++ src/constitutive/hybrid/elastic/multiplicative/mod.rs
@@ -162,10 +162,10 @@ impl<'a, C1: Elastic<'a>, C2: Elastic<'a>> MultiplicativeTrait for Multiplicativ
                 steps += 1;
             }
             if residual_norm >= ABS_TOL && steps == steps_maximum {
-                // return Err(ConstitutiveError::SolveError);
                 panic!("MAX STEPS REACHED")
+            } else {
+                Ok((deformation_gradient_1, deformation_gradient_2))
             }
-            Ok((deformation_gradient_1, deformation_gradient_2))
         }
     }
 }
